Updating based on b3694f8ae

+++ resource-manager/managementgroups/2023-04-01/managementgroups/model_createmanagementgroupdetails.go
@@ -13,7 +13,7 @@ type CreateManagementGroupDetails struct {
 	Parent      *CreateParentGroupInfo `json:"parent,omitempty"`
 	UpdatedBy   *string                `json:"updatedBy,omitempty"`
 	UpdatedTime *string                `json:"updatedTime,omitempty"`
-	Version     *float64               `json:"version,omitempty"`
+	Version     *int64                 `json:"version,omitempty"`
 }
 
 func (o *CreateManagementGroupDetails) GetUpdatedTimeAsTime() (*time.Time, error) {
